requirements.txt
ipython
fastapi
uvicorn
elasticsearch==7.13.4
sqlalchemy
pydantic
requests-aws4auth
logging
python-dotenv
boto3
awscli
pytest
pytest-asyncio
httpx
pyyaml
pytest-mock
botocore
